flytt mapping-logikk fra routes til service
